Stop trying to load /script/ui.xml
I think this feature was used way back in the very early days of
TWRP 2.0 on the Kindle Fire and has not been used since. I see no
reason to keep it around for now and we can always reinstate it
later if for some reason we need this feature again.

Change-Id: Ie9f3e5db7190732f813ae1e152aab2d85a4ca98c